refactor: rename route record to records
